+++ cvat-ui/src/components/annotation-page/standard-workspace/canvas-context-menu.tsx
@@ -28,7 +28,11 @@ export default function CanvasContextMenu(props: Props): JSX.Element | null {
 
     return ReactDOM.createPortal(
         <div className='cvat-canvas-context-menu' style={{ top, left }}>
-            <ObjectItemContainer key={activatedStateID} clientID={activatedStateID} />
+            <ObjectItemContainer
+                key={activatedStateID}
+                clientID={activatedStateID}
+                initialCollapsed
+            />
         </div>,
         window.document.body,
     );